refactor: move mockedlogger impl to common-tests
